opspec is there, mockfrom is in querytest
