+++ functions/outputs/builtin_test.go
@@ -2,5 +2,5 @@ package outputs_test
 
 import (
 	// We need to import the builtins for the tests to work.
-	_ "github.com/influxdata/flux/builtin"
+	_ "github.com/influxdata/flux/querytest/builtin"
 )
