--- querytest/builtin/doc.go
@@ -1,4 +1,0 @@
-// Package querytest.builtin contains the definition and initialization of testing-specific
-// operations and rewriting rules. This should be imported by tests that require end-to-end
-// query capabilities. This should be imported only from testing packages.
-package builtin
