+++ querytest/functions/doc.go
@@ -1,0 +1,2 @@
+// Package querytest.functions contains testing-specific functions.
+package functions
